Convert class to methods
